requirements.txt
altgraph==0.17.4
asgiref==3.3.1
blinker==1.8.2
bottle==0.13.2
click==8.1.7
colorama==0.4.6
Django==3.1.5
Flask==3.0.3
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysqlclient==2.2.4
packaging==24.2
pillow==10.4.0
proxy_tools==0.1.0
pycairo==1.27.0
PyGObject==3.50.0
pyinstaller-hooks-contrib==2024.11
pypng==0.0.20
PyQRCode==1.2.1
pytz==2020.5
pywebview==5.3.2
qrcode==7.4.2
QtPy==2.4.2
setuptools==75.8.0
SQLAlchemy==2.0.31
sqlparse==0.4.1
typing_extensions==4.12.2
tzdata==2024.1
Werkzeug==3.0.3
wheel==0.45.1
